types: refactor
